+++ src/pages/calendar/index.tsx
@@ -1,30 +1,16 @@
 import { Center, Flex, Heading, Text, VStack } from '@chakra-ui/react';
-import { useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router';
+import { useParams } from 'react-router';
 import { useSearchParams } from 'react-router-dom';
 
 import { Content } from '../../app/index';
-import { getCurrentTerm } from '../../app/shared/utils/terms';
 import { Term } from '../../shared/fetchers';
 import { SidebarTemplate } from '../../shared/SidebarTemplate';
-import { useSessionStorage } from '../../shared/useStorage';
 
 export function Calendar(): JSX.Element {
   const { term } = useParams();
   const [searchParams] = useSearchParams();
-  const [savedTerm, setSavedTerm] = useSessionStorage('user:term', getCurrentTerm());
-
-  const navigate = useNavigate();
 
   const pid = searchParams.get('pid');
-
-  useEffect(() => {
-    if (term) {
-      setSavedTerm(term);
-    } else {
-      navigate(`/calendar/${savedTerm}`, { replace: true });
-    }
-  }, [navigate, term, savedTerm, setSavedTerm]);
 
   return (
     <SidebarTemplate route="calendar" term={term as Term}>
